moved zone control away from main app.js
