+++ public/views/index.ejs
@@ -2,10 +2,7 @@
 <head>
     <link rel="stylesheet" href="static/css/styles.css">
 </head>
-
 <body>
-
-
     <div class="player2">
         <div class="playing-field">
             <div class="row-2">
